The idea of table creation logic takes shape.
